# base64: allow skipping padding = characters