--- a/lib/stdlib/test/base64_SUITE.erl
+++ b/lib/stdlib/test/base64_SUITE.erl
@@ -26,8 +26,8 @@
 -export([all/0, suite/0, groups/0, group/1]).
 
 %% Test cases must be exported.
--export([base64_encode/1, base64_encode_modes/1,
-	 base64_decode/1, base64_decode_modes/1,
+-export([base64_encode/1, base64_encode_to_string/1, base64_encode_modes/1,
+	 base64_decode/1, base64_decode_to_string/1, base64_decode_modes/1,
 	 base64_otp_5635/1, base64_otp_6279/1, big/1, illegal/1,
 	 mime_decode/1, mime_decode_modes/1,
 	 mime_decode_to_string/1, mime_decode_to_string_modes/1,
@@ -42,8 +42,8 @@ suite() ->
      {timetrap,{minutes,4}}].
 
 all() ->
-    [base64_encode, base64_encode_modes,
-     base64_decode, base64_decode_modes,
+    [base64_encode, base64_encode_to_string, base64_encode_modes,
+     base64_decode, base64_decode_to_string, base64_decode_modes,
      base64_otp_5635, base64_otp_6279, big, illegal,
      mime_decode, mime_decode_modes,
      mime_decode_to_string, mime_decode_to_string_modes,
@@ -58,76 +58,176 @@ group(roundtrip) ->
     [{timetrap,{minutes,10}}].
 
 %%-------------------------------------------------------------------------
-%% Test base64:encode/1.
+%% Test base64:encode/1,2.
 base64_encode(Config) when is_list(Config) ->
     %% Two pads
     <<"QWxhZGRpbjpvcGVuIHNlc2FtZQ==">> =
 	base64:encode("Aladdin:open sesame"),
+    <<"QWxhZGRpbjpvcGVuIHNlc2FtZQ">> =
+        base64:encode("Aladdin:open sesame", #{padding => false}),
     %% One pad
     <<"SGVsbG8gV29ybGQ=">> = base64:encode(<<"Hello World">>),
+    <<"SGVsbG8gV29ybGQ">> = base64:encode(<<"Hello World">>, #{padding => false}),
+    %% No pad
+    <<"QWxhZGRpbjpvcGVuIHNlc2Ft">> = base64:encode("Aladdin:open sesam"),
+    <<"QWxhZGRpbjpvcGVuIHNlc2Ft">> =
+        base64:encode("Aladdin:open sesam", #{padding => false}),
+
+    <<"MDEyMzQ1Njc4OSFAIzBeJiooKTs6PD4sLiBbXXt9">> =
+	base64:encode(<<"0123456789!@#0^&*();:<>,. []{}">>),
+    ok.
+
+%%-------------------------------------------------------------------------
+%% Test base64:encode_to_string/1,2.
+base64_encode_to_string(Config) when is_list(Config) ->
+    %% Two pads
+    "QWxhZGRpbjpvcGVuIHNlc2FtZQ==" =
+        base64:encode_to_string("Aladdin:open sesame"),
+    "QWxhZGRpbjpvcGVuIHNlc2FtZQ" =
+        base64:encode_to_string("Aladdin:open sesame", #{padding => false}),
+    %% One pad
+    "SGVsbG8gV29ybGQ=" = base64:encode_to_string(<<"Hello World">>),
+    "SGVsbG8gV29ybGQ" =
+        base64:encode_to_string(<<"Hello World">>, #{padding => false}),
     %% No pad
     "QWxhZGRpbjpvcGVuIHNlc2Ft" =
 	base64:encode_to_string("Aladdin:open sesam"),
+    "QWxhZGRpbjpvcGVuIHNlc2Ft" =
+        base64:encode_to_string("Aladdin:open sesam", #{padding => false}),
 
     "MDEyMzQ1Njc4OSFAIzBeJiooKTs6PD4sLiBbXXt9" =
 	base64:encode_to_string(<<"0123456789!@#0^&*();:<>,. []{}">>),
     ok.
 
 %%-------------------------------------------------------------------------
-%% Test base64:encode/2.
+%% Test base64:encode/2 and base64:encode_to_string/2.
 base64_encode_modes(Config) when is_list(Config) ->
     Data = <<23, 234, 63, 163, 239, 129, 253, 175, 171>>,
 
-    <<"F+o/o++B/a+r">> = base64:encode(Data, standard),
-    <<"F-o_o--B_a-r">> = base64:encode(Data, urlsafe),
+    <<"F+o/o++B/a+r">> = base64:encode(Data, #{mode => standard}),
+    <<"F-o_o--B_a-r">> = base64:encode(Data, #{mode => urlsafe}),
 
-    "F+o/o++B/a+r" = base64:encode_to_string(Data, standard),
-    "F-o_o--B_a-r" = base64:encode_to_string(Data, urlsafe),
+    "F+o/o++B/a+r" = base64:encode_to_string(Data, #{mode => standard}),
+    "F-o_o--B_a-r" = base64:encode_to_string(Data, #{mode => urlsafe}),
 
     ok.
 
 %%-------------------------------------------------------------------------
-%% Test base64:decode/1.
+%% Test base64:decode/1,2.
 base64_decode(Config) when is_list(Config) ->
     %% Two pads
     <<"Aladdin:open sesame">> =
-	base64:decode("QWxhZGRpbjpvcGVuIHNlc2FtZQ=="),
+        base64:decode(<<"QWxhZGRpbjpvcGVuIHNlc2FtZQ==">>),
+    <<"Aladdin:open sesame">> =
+        base64:decode("QWxhZGRpbjpvcGVuIHNlc2FtZQ=="),
+    <<"Aladdin:open sesame">> =
+        base64:decode(<<"QWxhZGRpbjpvcGVuIHNlc2FtZQ">>, #{padding => false}),
+    <<"Aladdin:open sesame">> =
+        base64:decode("QWxhZGRpbjpvcGVuIHNlc2FtZQ", #{padding => false}),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode(<<"QWxhZGRpbjpvcGVuIHNlc2FtZQ">>),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode("QWxhZGRpbjpvcGVuIHNlc2FtZQ"),
     %% One pad
     <<"Hello World">> = base64:decode(<<"SGVsbG8gV29ybGQ=">>),
+    <<"Hello World">> = base64:decode("SGVsbG8gV29ybGQ="),
+    <<"Hello World">> = base64:decode(<<"SGVsbG8gV29ybGQ">>, #{padding => false}),
+    <<"Hello World">> = base64:decode("SGVsbG8gV29ybGQ", #{padding => false}),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode_to_string(<<"SGVsbG8gV29ybGQ">>),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode_to_string("SGVsbG8gV29ybGQ"),
     %% No pad
     <<"Aladdin:open sesam">> =
-	base64:decode("QWxhZGRpbjpvcGVuIHNlc2Ft"),
+        base64:decode(<<"QWxhZGRpbjpvcGVuIHNlc2Ft">>),
+    <<"Aladdin:open sesam">> =
+        base64:decode("QWxhZGRpbjpvcGVuIHNlc2Ft"),
+    <<"Aladdin:open sesam">> =
+        base64:decode(<<"QWxhZGRpbjpvcGVuIHNlc2Ft">>, #{padding => false}),
+    <<"Aladdin:open sesam">> =
+        base64:decode("QWxhZGRpbjpvcGVuIHNlc2Ft", #{padding => false}),
 
     Alphabet = list_to_binary(lists:seq(0, 255)),
     Alphabet = base64:decode(base64:encode(Alphabet)),
+    Alphabet = base64:decode(base64:encode_to_string(Alphabet)),
+
+    %% Encoded base 64 strings may be divided by non base 64 chars.
+    %% In this cases whitespaces.
+    <<"0123456789!@#0^&*();:<>,. []{}">> =
+        base64:decode(
+            "MDEy MzQ1Njc4 \tOSFAIzBeJ \niooKTs6 PD4sLi \r\nBbXXt9"),
+    <<"0123456789!@#0^&*();:<>,. []{}">> =
+        base64:decode(
+            <<"MDEy MzQ1Njc4 \tOSFAIzBeJ \niooKTs6 PD4sLi \r\nBbXXt9">>),
+    ok.
+
+%%-------------------------------------------------------------------------
+%% Test base64:decode_to_string/1,2.
+base64_decode_to_string(Config) when is_list(Config) ->
+    %% Two pads
+    "Aladdin:open sesame" =
+        base64:decode_to_string(<<"QWxhZGRpbjpvcGVuIHNlc2FtZQ==">>),
+    "Aladdin:open sesame" =
+        base64:decode_to_string("QWxhZGRpbjpvcGVuIHNlc2FtZQ=="),
+    "Aladdin:open sesame" =
+        base64:decode_to_string(<<"QWxhZGRpbjpvcGVuIHNlc2FtZQ">>, #{padding => false}),
+    "Aladdin:open sesame" =
+        base64:decode_to_string("QWxhZGRpbjpvcGVuIHNlc2FtZQ", #{padding => false}),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode_to_string(<<"QWxhZGRpbjpvcGVuIHNlc2FtZQ">>),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode_to_string("QWxhZGRpbjpvcGVuIHNlc2FtZQ"),
+    %% One pad
+    "Hello World" = base64:decode_to_string(<<"SGVsbG8gV29ybGQ=">>),
+    "Hello World" = base64:decode_to_string("SGVsbG8gV29ybGQ="),
+    "Hello World" = base64:decode_to_string(<<"SGVsbG8gV29ybGQ">>, #{padding => false}),
+    "Hello World" = base64:decode_to_string("SGVsbG8gV29ybGQ", #{padding => false}),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode_to_string(<<"SGVsbG8gV29ybGQ">>),
+    {'EXIT', {'missing_padding', _}} =
+        catch base64:decode_to_string("SGVsbG8gV29ybGQ"),
+    %% No pad
+    "Aladdin:open sesam" =
+        base64:decode_to_string(<<"QWxhZGRpbjpvcGVuIHNlc2Ft">>),
+    "Aladdin:open sesam" =
+        base64:decode_to_string("QWxhZGRpbjpvcGVuIHNlc2Ft"),
+    "Aladdin:open sesam" =
+        base64:decode_to_string(<<"QWxhZGRpbjpvcGVuIHNlc2Ft">>, #{padding => false}),
+    "Aladdin:open sesam" =
+        base64:decode_to_string("QWxhZGRpbjpvcGVuIHNlc2Ft", #{padding => false}),
+
+    Alphabet = lists:seq(0, 255),
+    Alphabet = base64:decode_to_string(base64:encode(Alphabet)),
+    Alphabet = base64:decode_to_string(base64:encode_to_string(Alphabet)),
 
     %% Encoded base 64 strings may be divided by non base 64 chars.
     %% In this cases whitespaces.
     "0123456789!@#0^&*();:<>,. []{}" =
-	base64:decode_to_string(
-	  "MDEy MzQ1Njc4 \tOSFAIzBeJ \niooKTs6 PD4sLi \r\nBbXXt9"),
+        base64:decode_to_string(
+            "MDEy MzQ1Njc4 \tOSFAIzBeJ \niooKTs6 PD4sLi \r\nBbXXt9"),
     "0123456789!@#0^&*();:<>,. []{}" =
-	base64:decode_to_string(
-	  <<"MDEy MzQ1Njc4 \tOSFAIzBeJ \niooKTs6 PD4sLi \r\nBbXXt9">>),
+        base64:decode_to_string(
+            <<"MDEy MzQ1Njc4 \tOSFAIzBeJ \niooKTs6 PD4sLi \r\nBbXXt9">>),
     ok.
 
 %%-------------------------------------------------------------------------
-%% Test base64:decode/2.
+%% Test base64:decode/2 and base64:decode_to_string/2
 base64_decode_modes(Config) when is_list(Config) ->
     DataBin = <<23, 234, 63, 163, 239, 129, 253, 175, 171>>,
     DataStr = [23, 234, 63, 163, 239, 129, 253, 175, 171],
 
-    DataBin = base64:decode("F+o/o++B/a+r", standard),
-    DataBin = base64:decode("F-o_o--B_a-r", urlsafe),
-    {'EXIT', _} = catch base64:decode("F-o_o--B_a-r", standard),
-    {'EXIT', _} = catch base64:decode("F+o/o++B/a+r", urlsafe),
+    DataBin = base64:decode("F+o/o++B/a+r", #{mode => standard}),
+    DataBin = base64:decode("F-o_o--B_a-r", #{mode => urlsafe}),
+    {'EXIT', _} = catch base64:decode("F-o_o--B_a-r", #{mode => standard}),
+    {'EXIT', _} = catch base64:decode("F+o/o++B/a+r", #{mode => urlsafe}),
 
-    DataStr = base64:decode_to_string("F+o/o++B/a+r", standard),
-    DataStr = base64:decode_to_string("F-o_o--B_a-r", urlsafe),
-    {'EXIT', _} = catch base64:decode_to_string("F-o_o--B_a-r", standard),
-    {'EXIT', _} = catch base64:decode_to_string("F+o/o++B/a+r", urlsafe),
+    DataStr = base64:decode_to_string("F+o/o++B/a+r", #{mode => standard}),
+    DataStr = base64:decode_to_string("F-o_o--B_a-r", #{mode => urlsafe}),
+    {'EXIT', _} = catch base64:decode_to_string("F-o_o--B_a-r", #{mode => standard}),
+    {'EXIT', _} = catch base64:decode_to_string("F+o/o++B/a+r", #{mode => urlsafe}),
 
     ok.
+
 %%-------------------------------------------------------------------------
 %% OTP-5635: Some data doesn't pass through base64:decode/1 correctly.
 base64_otp_5635(Config) when is_list(Config) ->
@@ -210,26 +310,26 @@ mime_decode(Config) when is_list(Config) ->
 
 %% Test base64:mime_decode/2.
 mime_decode_modes(Config) when is_list(Config) ->
-    MimeDecode = fun (In, Mode) ->
-                                Out = base64:mime_decode(In, Mode),
-                                Out = base64:mime_decode(binary_to_list(In), Mode)
+    MimeDecode = fun (In, Options) ->
+                                Out = base64:mime_decode(In, Options),
+                                Out = base64:mime_decode(binary_to_list(In), Options)
                  end,
 
     %% The following all decode to the same data.
     Data = <<23, 234, 63, 163, 239, 129, 253, 175, 171>>,
-    Data = MimeDecode(<<"F+o/o++B/a+r">>, standard),
-    Data = MimeDecode(<<"F-o_o--B_a-r">>, urlsafe),
+    Data = MimeDecode(<<"F+o/o++B/a+r">>, #{mode => standard}),
+    Data = MimeDecode(<<"F-o_o--B_a-r">>, #{mode => urlsafe}),
 
     %% The following decodes to different data depending on mode.
     Base64 = <<"AB+C+D/E/FG-H-I_J_KL">>,
     %% In standard mode, "-" and "_" are invalid and thus ignored.
     %% The base64 string to be decoded is equivalent to "AB+C+D/E/FGHIJKL".
     <<0, 31, 130, 248, 63, 196, 252, 81, 135, 32, 146, 139>> =
-        MimeDecode(Base64, standard),
+        MimeDecode(Base64, #{mode => standard}),
     %% In urlsafe mode, "+" and "/" are invalid and thus ignored.
     %% The base64 string to be decoded is equivalent to "ABCDEFG-H-I_J_KL".
     <<0, 16, 131, 16, 81, 190, 31, 226, 63, 39, 242, 139>> =
-        MimeDecode(Base64, urlsafe),
+        MimeDecode(Base64, #{mode => urlsafe}),
 
     ok.
 
@@ -286,26 +386,26 @@ mime_decode_to_string(Config) when is_list(Config) ->
 
 %% Test base64:mime_decode_to_string/2.
 mime_decode_to_string_modes(Config) when is_list(Config) ->
-    MimeDecode = fun (In, Mode) ->
-                                Out = base64:mime_decode_to_string(In, Mode),
-                                Out = base64:mime_decode_to_string(binary_to_list(In), Mode)
+    MimeDecode = fun (In, Options) ->
+                                Out = base64:mime_decode_to_string(In, Options),
+                                Out = base64:mime_decode_to_string(binary_to_list(In), Options)
                  end,
 
     %% The following all decode to the same data.
     Data = [23, 234, 63, 163, 239, 129, 253, 175, 171],
-    Data = MimeDecode(<<"F+o/o++B/a+r">>, standard),
-    Data = MimeDecode(<<"F-o_o--B_a-r">>, urlsafe),
+    Data = MimeDecode(<<"F+o/o++B/a+r">>, #{mode => standard}),
+    Data = MimeDecode(<<"F-o_o--B_a-r">>, #{mode => urlsafe}),
 
     %% The following decodes to different data depending on mode.
     Base64 = <<"AB+C+D/E/FG-H-I_J_KL">>,
     %% In standard mode, "-" and "_" are invalid and thus ignored.
     %% The base64 string to be decoded is equivalent to "AB+C+D/E/FGHIJKL".
     [0, 31, 130, 248, 63, 196, 252, 81, 135, 32, 146, 139] =
-        MimeDecode(Base64, standard),
+        MimeDecode(Base64, #{mode => standard}),
     %% In urlsafe mode, "+" and "/" are invalid and thus ignored.
     %% The base64 string to be decoded is equivalent to "ABCDEFG-H-I_J_KL".
     [0, 16, 131, 16, 81, 190, 31, 226, 63, 39, 242, 139] =
-        MimeDecode(Base64, urlsafe),
+        MimeDecode(Base64, #{mode => urlsafe}),
 
     ok.