@bjorng 2023-01-25
> This amends the interface in a backwards-incompatible way, since the last change wasn't released yet, so should have no expectation of backwards-compatibility.

Can you clarify this? Is it compatible with OTP 25 and previous releases or not?