@michalmuskala 2023-01-25
It is fully compatible with OTP 25 - there only the /1 functions exists and their behaviour didn't change. However, it is not compatible with current master, but this new interface was never released