@bjorng 2023-01-25
Have you done any benchmarking to make sure that there is no a significant slowdown because of the new options?