@bjorng 2023-01-26
Here is a benchmark I have been using to test `base64` performance:

https://gist.github.com/bjorng/522470f313ac01d16156a597657f97bb

It is based on the benchmark in #5639.