@max-au 2023-01-27
Quick test on my machine shows no regression at all, at least for small strings. Here is before this PR:
```
./erlperf 'base64:decode(<<"SGVsbG8gV29ybGQ=">>).' -l 10M -s 10
Code                                       ||   Samples       Avg   StdDev    Median      P99  Iteration
base64:decode(<<"SGVsbG8gV29ybGQ=">>).      1        10   1686 ms    1.08%   1672 ms  1710 ms     168 ns
```

And this is after, using new option:
```
./erlperf 'base64:decode(<<"SGVsbG8gV29ybGQ">>, #{padding => false}).' -l 10M -s 10
Code                                                           ||   Samples       Avg   StdDev    Median      P99  Iteration
base64:decode(<<"SGVsbG8gV29ybGQ">>, #{padding => false}).      1        10   1654 ms    0.28%   1656 ms  1659 ms     165 ns
```

I also tried with random 1Mb input, with the same effect.

I wonder which case is expected to give the worst performance, when padding isn't needed, or when it needs 3 `=` characters?