@michalmuskala 2023-01-27
I couldn't measure any significant peformance difference on my mac. 

Before:
```
== Testing with 100 B ==
fun base64:encode/1: 1000000 iterations in 1285 ms: 778210 it/sec
fun base64:decode/1: 1000000 iterations in 1600 ms: 625000 it/sec

== Testing with 1 MB ==
fun base64:encode/1: 1000 iterations in 12767 ms: 78 it/sec
fun base64:decode/1: 1000 iterations in 15620 ms: 64 it/sec
```

After:
```
== Testing with 100 B ==
fun base64:encode/1: 1000000 iterations in 1280 ms: 781250 it/sec
fun base64:decode/1: 1000000 iterations in 1594 ms: 627352 it/sec

== Testing with 1 MB ==
fun base64:encode/1: 1000 iterations in 12671 ms: 78 it/sec
fun base64:decode/1: 1000 iterations in 15469 ms: 64 it/sec
```

@max-au the amount of padding shouldn't impact performance almost at all - it's basically linear in the length of the string. Since there's no copying or anything like that, it's just traversing or not the extra couple padding bytes - it shouldn't be noticeable.