@bjorng 2023-01-27
Thanks for benchmarking, @max-au and @michalmuskala.

The OTP Technical Board will have to approve any API changes. To be able to approve the API, the board members will need to see a description of the API (not by reading the code). I have looked at your code and based on that look it seems to me that the API is reasonable and I find it likely that it will be approved. You can either write the documentation now or write a short API description in a comment in this PR.