@bjorng 2023-02-02
I have pushed two fixup commits. The first is addition of links (suggested by @RaimoNiskanen); the other is an updated description of the default value that I find less confusing. If you agree, please squash the commits. When you are done, I'll merge the PR.